make uint and int implementations more symmetric
