+++ gadgets/benches/integer_arithmetic.rs
@@ -16,6 +16,8 @@
 
 use snarkvm_gadgets::traits::utilities::{
     alloc::AllocGadget,
+    arithmetic::*,
+    num::Number,
     uint::{UInt, UInt128, UInt16, UInt32, UInt64, UInt8},
 };
 use snarkvm_r1cs::{ConstraintSystem, Fr, TestConstraintSystem};
