+++ gadgets/src/algorithms/commitment/blake2s.rs
@@ -20,7 +20,8 @@ use crate::{
         algorithms::CommitmentGadget,
         utilities::{
             alloc::AllocGadget,
-            uint::unsigned_integer::{UInt, UInt8},
+            uint::unsigned_integer::UInt8,
+            num::Number,
             ToBytesGadget,
         },
     },
