+++ gadgets/src/algorithms/prf/tests.rs
@@ -22,7 +22,8 @@ use crate::{
             alloc::AllocGadget,
             boolean::{AllocatedBit, Boolean},
             eq::EqGadget,
-            uint::unsigned_integer::{UInt, UInt8},
+            uint::unsigned_integer::UInt8,
+            num::Number,
         },
     },
 };
