+++ gadgets/src/fields/fp.rs
@@ -21,7 +21,8 @@ use crate::{
         boolean::{AllocatedBit, Boolean},
         eq::{ConditionalEqGadget, EqGadget, NEqGadget},
         select::{CondSelectGadget, ThreeBitCondNegLookupGadget, TwoBitLookupGadget},
-        uint::unsigned_integer::{UInt, UInt8},
+        uint::unsigned_integer::UInt8,
+        num::Number,
         ToBitsGadget,
         ToBytesGadget,
     },
