+++ gadgets/src/traits/utilities/uint/arithmetic/sub.rs
@@ -14,17 +14,31 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
-#[macro_use]
-mod macros;
+use snarkvm_fields::PrimeField;
 
-pub mod unsigned_integer;
-pub use unsigned_integer::*;
+use crate::{
+    utilities::{
+        arithmetic::Sub,
+        uint::*,
+    },
+};
+use snarkvm_r1cs::{errors::SynthesisError, ConstraintSystem};
 
-pub mod uint128;
-pub use uint128::*;
+macro_rules! sub_uint_impl {
+    ($($gadget: ident)*) => ($(
+        impl Sub for $gadget {
+            type ErrorType = SynthesisError;
 
-pub mod arithmetic;
-pub mod relational;
+            fn sub<F: PrimeField, CS: ConstraintSystem<F>>(&self, mut cs: CS, other: &Self) -> Result<Self, Self::ErrorType> {
+                // pseudocode:
+                //
+                // a - b
+                // a + (-b)
 
-#[cfg(test)]
-mod tests;
+                Self::addmany(&mut cs.ns(|| "add_not"), &[self.clone(), other.negate()])
+            }
+        }
+    )*)
+}
+
+sub_uint_impl!(UInt8 UInt16 UInt32 UInt64 UInt128);
