+++ gadgets/src/traits/utilities/uint/tests/uint16.rs
@@ -18,6 +18,8 @@ use crate::utilities::{
     alloc::AllocGadget,
     boolean::Boolean,
     uint::{UInt, UInt16},
+    arithmetic::*,
+    num::Number,
 };
 use snarkvm_fields::{One, Zero};
 use snarkvm_r1cs::{ConstraintSystem, Fr, TestConstraintSystem};
