+++ gadgets/src/traits/utilities/uint/tests/uint64.rs
@@ -14,10 +14,14 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkVM library. If not, see <https://www.gnu.org/licenses/>.
 
-use crate::utilities::{
-    alloc::AllocGadget,
-    boolean::Boolean,
-    uint::{UInt, UInt64},
+use crate::{
+    utilities::{
+        alloc::AllocGadget,
+        arithmetic::*,
+        boolean::Boolean,
+        num::Number,
+        uint::{UInt, UInt64},
+    },
 };
 use snarkvm_fields::{One, Zero};
 use snarkvm_r1cs::{ConstraintSystem, Fr, TestConstraintSystem};
